2 linker errors resolved
